Backend/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
werkzeug
itsdangerous
opencv-python
face-recognition
openpyxl